+++ src/app/api/models/unit.ts
@@ -34,6 +34,7 @@ import {HttpClient, HttpParams} from '@angular/common/http';
 import {TaskPrerequisiteService} from '../services/task-prerequisite.service';
 import {MarkingSession} from './marking-session';
 import {MarkingSessionService} from '../services/marking-session.service';
+import {TaskPrerequisite} from './task-prerequisite';
 
 export class Unit extends Entity {
   id: number;
@@ -731,5 +732,14 @@ export class Unit extends Entity {
       `${AppInjector.get(DoubtfireConstants).API_URL}/csv/units/${this.id}/staff_notes`,
       `${this.name}-StaffNotes.csv`,
     );
+  }
+
+  public get taskDefinitionsPrerequisitesUrl(): string {
+    return `${AppInjector.get(DoubtfireConstants).API_URL}/units/${this.id}/task_prerequisites`;
+  }
+
+  public getTaskPrerequisites(): Observable<TaskPrerequisite[]> {
+    const prerequisiteService = AppInjector.get(TaskPrerequisiteService);
+    return prerequisiteService.getUnitPrerequisites(this.id);
   }
 }